Add Monitor documentation
