+++ Sources/Monitor/Interface.swift
@@ -7,9 +7,18 @@ import Foundation
 import Combine
 import A11yFeature
 
+/// Monitors updates to accessibility features that are currently being observed and provides an interface to query the current status of a given accessibility feature.
 public class Monitor {
 
+    /// A list of accessibility features that are currently being observed.
     public let featuresSubject: CurrentValueSubject<[A11yFeature], Never>
+
+    /// Provides the ability to query whether or not a particular accessibility feature is currently enabled or not, by its type.
+    ///
+    /// - Parameter featureType: The type of accessibility to query whether or not it is currently enabled.
+    ///
+    /// - Returns: A boolean value that indicates whether or not the accessibility feature is currently enabled.
+    ///
     public let isFeatureEnabled: (_ feature: A11yFeatureType) -> Bool
 
     public required init(
